Move TaskBloc to Bloc folder
