+++ lib/bloc/task_state.dart
