+++ lib/screens/pending_tasks_screen.dart
@@ -1,8 +1,8 @@
-import 'package:bloc_finals_exam/cubit/task_bloc.dart';
+import 'package:bloc_finals_exam/bloc/task_bloc.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../cubit/task_state.dart';
+import '../bloc/task_state.dart';
 import '../models/task.dart';
 import '../test_data.dart';
 import '../widgets/tasks_list.dart';
